Compatability with .NET 2.5 and Unity: - Added some default return value where a default case in a switch wasn't enough - Replace decimal with double

+++ Cottle/src/Builtins/BuiltinOperators.cs
@@ -21,7 +21,7 @@ namespace Cottle.Builtins
 
 		public static readonly IFunction	operatorDiv = new NativeFunction ((values) =>
 		{
-			decimal denominator;
+			double denominator;
 
 			denominator = values[1].AsNumber;
 
@@ -54,7 +54,7 @@ namespace Cottle.Builtins
 
 		public static readonly IFunction	operatorMod = new NativeFunction ((values) =>
 		{
-			decimal denominator;
+			double denominator;
 
 			denominator = values[1].AsNumber;
